requirements.txt
tensorflow
opencv-python
PyQt5
python-vlc
numpy
